OS Portability Improvement
Rather than constructing input and output file paths from user-given text strings, this updated version takes advantage of the Python package "pathlib" for processing OS-agnostic file paths for reading and writing.  It also automatically creates the output file and names it results.txt.  Note that if a file with this name already exists in the folder specified, it will be overwritten.

The pathlib functions used here work best in Python 3.6.  Tests run in Python 3.5 produced errors when trying to use the agnostic path object with more common Python functions such as listdir() and open().